# Conflicts:
#	src/main/resources/static/ManageResignType.html
#	src/main/resources/static/attendanceClassManager.html
#	src/main/resources/static/attendanceDailyStatistics.html
#	src/main/resources/static/attendanceGroupManager.html
#	src/main/resources/static/attendanceManagement.html
#	src/main/resources/static/attendanceMonthSummary.html
#	src/main/resources/static/attendanceOriginalRecord.html
#	src/main/resources/static/attendancePunchTime.html
#	src/main/resources/static/css/ManageResignType.css
#	src/main/resources/static/css/salaryReport.css
#	src/main/resources/static/js/ManageResignType.js
#	src/main/resources/static/js/attendanceClassManager.js
#	src/main/resources/static/js/attendanceGroupManager.js
#	src/main/resources/static/js/sample.js
#	src/main/resources/static/salaryReport.html

+++ src/main/java/com/xplusplus/security/service/ProjectUserService.java
@@ -214,14 +214,25 @@ public class ProjectUserService {
      */
 	@Transactional
 	public int assignUsersToProject(Long projectId, String[] userIds) {
+
+		if(userIds == null){
+            throw new SecurityExceptions(EnumExceptions.FAILED_NOT_USERIDS);
+		}
+
 	    // 验证项目是否存在
         Project project = projectRepository.findOne(projectId);
         if(project == null){
             throw new SecurityExceptions(EnumExceptions.ASSIGN_FAILED_PROJECT_NOT_EXIST);
         }
 
-        // 删除已分配的记录
-        projectUserRepository.deleteByProject(project);
+        while(true) {
+            // 删除已分配的记录
+            projectUserRepository.deleteByProject(project);
+            List<ProjectUser> projectUsers = projectUserRepository.findByProject(project);
+            if(projectUsers == null || projectUsers.size() == 0){
+                break;
+            }
+        }
 
         // 用set去重
         HashSet<String> userIdSet = new HashSet<>(Arrays.asList(userIds));
